// program/clients/typescript/src/generated/types/feeType.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import { combineCodec, getEnumDecoder, getEnumEncoder, type FixedSizeCodec, type FixedSizeDecoder, type FixedSizeEncoder } from '@solana/kit';



export enum FeeType { Bps, Fixed };

export type FeeTypeArgs = FeeType;




export function getFeeTypeEncoder(): FixedSizeEncoder<FeeTypeArgs> {
  return getEnumEncoder(FeeType);
}



export function getFeeTypeDecoder(): FixedSizeDecoder<FeeType> {
  return getEnumDecoder(FeeType);
}




export function getFeeTypeCodec(): FixedSizeCodec<FeeTypeArgs, FeeType> {
  return combineCodec(getFeeTypeEncoder(), getFeeTypeDecoder());
}